+++ main.py
@@ -1,148 +1,9 @@
-import argparse
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-from torchvision import datasets, transforms
-from torch.optim.lr_scheduler import StepLR
-from torch.utils.data import Subset
-import numpy as np
-
-class NeuralNet(nn.Module):
-    def __init__(self):
-        super(NeuralNet, self).__init__()
-        # Convolutional Layers
-        self.conv1 = nn.Conv2d(1, 32, kernel_size=3, padding=1)
-        self.pool1 = nn.MaxPool2d(2, 2)
-        
-        self.conv2 = nn.Conv2d(32, 64, kernel_size=3, padding=1)
-        self.pool2 = nn.MaxPool2d(2, 2)
-        
-        self.conv3 = nn.Conv2d(64, 128, kernel_size=3, padding=1)
-        self.pool3 = nn.MaxPool2d(2, 2)
-        
-        self.conv4 = nn.Conv2d(128, 256, kernel_size=3, padding=1)
-        self.pool4 = nn.MaxPool2d(2, 2)
-        
-        # Fully Connected Layer
-        self.fc1 = nn.Linear(256 * 1 * 1, 10)  # Adjusted for the output size after convolutions
-
-    def forward(self, x):
-        x = F.relu(self.conv1(x))
-        x = self.pool1(x)
-        
-        x = F.relu(self.conv2(x))
-        x = self.pool2(x)
-        
-        x = F.relu(self.conv3(x))
-        x = self.pool3(x)
-        
-        x = F.relu(self.conv4(x))
-        x = self.pool4(x)
-        
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        return F.log_softmax(x, dim=1)
-
-def train(model, device, train_loader, optimizer, epoch):
-    model.train()
-    for batch_idx, (data, target) in enumerate(train_loader):
-        data, target = data.to(device), target.to(device)
-        optimizer.zero_grad()
-        output = model(data)
-        loss = F.nll_loss(output, target)
-        loss.backward()
-        optimizer.step()
-        if batch_idx % 10 == 0:
-            print(f'Train Epoch: {epoch} [{batch_idx * len(data)}/{len(train_loader.dataset)} '
-                  f'({100. * batch_idx / len(train_loader):.0f}%)]\tLoss: {loss.item():.6f}')
-
-def test(model, device, test_loader):
-    model.eval()
-    test_loss = 0
-    correct = 0
-    with torch.no_grad():
-        for data, target in test_loader:
-            data, target = data.to(device), target.to(device)
-            output = model(data)
-            test_loss += F.nll_loss(output, target, reduction='sum').item()  # sum up batch loss
-            pred = output.argmax(dim=1, keepdim=True)  # get the index of the max log-probability
-            correct += pred.eq(target.view_as(pred)).sum().item()
-
-    test_loss /= len(test_loader.dataset)
-
-    print(f'\nTest set: Average loss: {test_loss:.4f}, '
-          f'Accuracy: {correct}/{len(test_loader.dataset)} '
-          f'({100. * correct / len(test_loader.dataset):.0f}%)\n')
-    
-    return 100. * correct / len(test_loader.dataset)
+from .CNN import Run_CNN
+from .Noise import Add_Noise
 
 def main():
-    # Training settings
-    parser = argparse.ArgumentParser(description='PyTorch MNIST CNN')
-    parser.add_argument('--batch-size', type=int, default=64,
-                        help='input batch size for training (default: 64)')
-    parser.add_argument('--epochs', type=int, default=14,
-                        help='number of epochs to train (default: 14)')
-    parser.add_argument('--lr', type=float, default=1.0,
-                        help='learning rate (default: 1.0)')
-    parser.add_argument('--data-fraction', type=float, default=1.0,
-                        help='Fraction of training data to use (default: 1.0)')
-    parser.add_argument('--seed', type=int, default=1,
-                        help='random seed (default: 1)')
-    parser.add_argument('--save-model', action='store_true', default=False,
-                        help='For Saving the current Model')
-    args = parser.parse_args()
-
-    torch.manual_seed(args.seed)
-
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    print(f'Using device: {device}')
-
-    train_kwargs = {'batch_size': args.batch_size}
-    test_kwargs = {'batch_size': 1000}
-    if torch.cuda.is_available():
-        train_kwargs.update({'num_workers': 1,
-                             'pin_memory': True,
-                             'shuffle': True})
-        test_kwargs.update({'num_workers': 1,
-                            'pin_memory': True,
-                            'shuffle': False})
-
-    transform = transforms.Compose([
-        transforms.ToTensor(),
-        transforms.Normalize((0.1307,), (0.3081,))
-    ])
-    dataset1 = datasets.MNIST('../data', train=True, download=True,
-                              transform=transform)
-    dataset2 = datasets.MNIST('../data', train=False,
-                              transform=transform)
-    
-    # Adjust the training dataset based on data_fraction
-    if args.data_fraction < 1.0:
-        num_train = int(len(dataset1) * args.data_fraction)
-        indices = np.random.choice(len(dataset1), num_train, replace=False)
-        subset = Subset(dataset1, indices)
-        train_loader = torch.utils.data.DataLoader(subset, **train_kwargs)
-        print(f'Using {num_train} training samples out of {len(dataset1)}')
-    else:
-        train_loader = torch.utils.data.DataLoader(dataset1, **train_kwargs)
-        print(f'Using all {len(dataset1)} training samples')
-
-    test_loader = torch.utils.data.DataLoader(dataset2, **test_kwargs)
-
-    model = NeuralNet().to(device)
-    optimizer = optim.Adadelta(model.parameters(), lr=args.lr)
-
-    scheduler = StepLR(optimizer, step_size=1, gamma=0.7)
-    for epoch in range(1, args.epochs + 1):
-        train(model, device, train_loader, optimizer, epoch)
-        test_accuracy = test(model, device, test_loader)
-        scheduler.step()
-
-    if args.save_model:
-        torch.save(model.state_dict(), "mnist_cnn.pt")
-        print("Model saved as mnist_cnn.pt")
+    Run_CNN()
+    Add_Noise()
 
 if __name__ == '__main__':
     main()
